Change assembly function signature to uint64_t[]
The assembly code doesn't care what the signature of the function is, so long as the first argument is a pointer to the start of the 1600 bits that make up the state. We can help cgo when calling the permute function and avoid having to copy data or cast anything as now the uint64[] in golang maps cleanly to a uint64_t[] (at least from cgo's perspective). This also means we don't have to copy the data from go to c before using it.