CMake build updates
Now we use generator expressions inside of the arm file so that we get the right location of the header files as well as the assembly library to link against for cgo.
Also enabling cgo always and setting the appropriate compiler for cgo to use with the CC environment variable. Additionally hard coding the GOARM GOARCH and GOOS variables for now to simplify compilation.